Add tabbar and simple story views
