+++ README.md
@@ -153,13 +153,13 @@
 
 ### 5.3 Сторис
 
-- [ ] tabbar, показывающий два основных экрана: поиск и настройки;
+- [x] tabbar, показывающий два основных экрана: поиск и настройки;
 - [ ] навигация типа master-detail между всеми экранами в дизайне;
 - [ ] сториз, которые открываются модально. 
 
 Self-check:
 
-- [ ] Создан tabbar для переключения между главным экраном поиска поездок и экраном настроек.
+- [x] Создан tabbar для переключения между главным экраном поиска поездок и экраном настроек.
 - [ ] В момент начала поиска организован полноэкранный показ выбора города с помощью `.fullScreenCover()`.
 - [ ] После выбора города реализован переход на экран выбора станции.
 - [ ] После выбора станции реализован возврат назад на главный экран.
