--- TravelSchedule/TabBar.swift
@@ -1,8 +1,0 @@
-//
-//  TabBar.swift
-//  TravelSchedule
-//
-//  Created by andrei.chenchik on 05.04.24.
-//
-
-import Foundation
